descriptors: correct 'pkh()' in the compatibility matrix
`pk_h()` is a Miniscript fragment (the equivalent of `pk_k()` for pubkey
hashes) which needs a `c:` in order to be a base expression.

The documentation was misleadingly implying that Bitcoin Core was using
`pkh()` to describe a `pk_h()` fragment, which is wrong (the former is
an alias of `c:pk_h()`).

The table also doesn't list `pk_k()`. Therefore just list the `pkh()`
and `pk()` aliases.